requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
PyPDF2>=3.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
googletrans>=4.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
pyaudio>=0.2.11
